refactor extension object in utils package
