+++ src/main/scala/co/topl/modifier/transaction/serialization/ArbitTransferSerializer.scala
@@ -5,7 +5,7 @@ import co.topl.attestation.serialization.{ProofSerializer, PropositionSerializer
 import co.topl.modifier.transaction.ArbitTransfer
 import co.topl.modifier.box.TokenValueHolder
 import co.topl.utils.Int128
-import co.topl.utils.serialization.Extensions._
+import co.topl.utils.Extensions._
 import co.topl.utils.serialization.{BifrostSerializer, Reader, Writer}
 
 import scala.language.existentials
