+++ src/main/scala/co/topl/network/peer/LocalAddressPeerFeatureSerializer.scala
@@ -2,7 +2,7 @@ package co.topl.network.peer
 
 import java.net.{InetAddress, InetSocketAddress}
 
-import co.topl.utils.serialization.Extensions._
+import co.topl.utils.Extensions._
 import co.topl.utils.serialization.{BifrostSerializer, Reader, Writer}
 
 object LocalAddressPeerFeatureSerializer extends BifrostSerializer[LocalAddressPeerFeature] {
